backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
python-magic-bin==0.4.14
httpx==0.27.2
matplotlib==3.9.2
seaborn==0.13.2
numpy==1.26.4
python-dotenv==1.0.1
